refactor: move syncfl files
While top-level horizontal files (top_aggregator.py,
middle_aggregator.py and trainer.py) are basis of other modes under
horizontal category, the directory structure is not well organized.

These files are now moved into horizontal/syncfl folder while we keep
backward compatibility.
